+++ nouvelles/forms.py
@@ -19,14 +19,15 @@ class ArchiveFiltersForm(forms.Form):
 
     criticality_choices += list(Article.CRITICALITY_CHOICES)
 
-    title = forms.CharField(label='Title', required=False)
+    q = forms.CharField(label='Search', required=False)
     author = UserModelChoiceField(
         label='Author',
         queryset=User.objects.all().order_by('username'),
         to_field_name='username',
         required=False,
         empty_label='-- All authors --')
-    date = forms.DateField(required=False)
+    date = forms.DateField(required=False,
+                           widget=forms.TextInput(attrs={'placeholder': 'YYYY-MM-DD'}))
     criticality = forms.ChoiceField(
         label='Criticality',
         required=False,